BacenSGS

Documentation for BacenSGS.

BacenSGS.parse_resp_timeseries

BacenSGS.search_timeseries

BacenSGS.timeseries_values

BacenSGS.parse_resp_timeseries — Method

function parse_resp_timeseries(resp)

Localiza campos a partir da resposta da query de times series

BacenSGS.search_timeseries — Method

function search_timeseries(query)

Localiza séries por query - código(Integer) ou texto(String)

BacenSGS.timeseries_values — Method

function timeseries_values(timeseries::Vector;dtInicial::Dates.Date=Date(2018),dtFinal=missing, progress=nothing)

Retorna os valores de um conjunto de timeseries a partir de uma data inicial dtInicial até a data final dtFinal